Updating test for new release & changed periodUnits & CalendarType
### Payroll NZ API
## PayRun.cs & PayRunCalendar.cs
- now refers to CalendarType class for enum options.
## Taxsettings.cs
- Periodunit type changed from int to decimal
Updated test project's RestSharp version from 106.6.7 to 106.11.4.
SDK version updated to 3.9.0

+++ Xero.NetStandard.OAuth2/Model/PayrollNz/TaxSettings.cs
@@ -69,7 +69,7 @@ namespace Xero.NetStandard.OAuth2.Model.PayrollNz
         /// </summary>
         /// <value>The number of units for the period type</value>
         [DataMember(Name="periodUnits", EmitDefaultValue=false)]
-        public int? PeriodUnits { get; set; }
+        public decimal? PeriodUnits { get; set; }
 
         /// <summary>
         /// Tax rate for STC and WT
